requirements.txt
appdirs==1.4.4
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
decorator==5.0.9
dogpile.cache==1.1.3
idna==2.10
iso8601==0.1.14
jmespath==0.10.0
jsonpatch==1.32
jsonpointer==2.1
keystoneauth1==4.3.1
munch==2.5.0
netifaces==0.11.0
openstacksdk==0.57.0
os-service-types==1.7.0
paramiko @ git+https://github.com/kkovaacs/paramiko.git@d43ed76a44c451b2b9031883a1a7e55de59fdf79 # adds support of modern RSA keys
pbr==5.6.0
pycparser==2.20
PyNaCl==1.4.0
PyYAML==5.4.1
requests==2.25.1
requestsexceptions==1.4.0
six==1.16.0
stevedore==3.3.0
urllib3==1.26.5
